First commit for Scene delegate code removal
