+++ lib/mayaHydra/hydraExtensions/mayaHydraParams.h
@@ -1,5 +1,5 @@
 //
-// Copyright 2019 Luma Pictures
+// Copyright 2024 Autodesk, Inc. All rights reserved.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -13,6 +13,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 //
+
 #ifndef MAYAHYDRALIB_PARAMS_H
 #define MAYAHYDRALIB_PARAMS_H
